requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
